requirements.txt
requests
beautifulsoup4
pandas
easyocr
python-dotenv
opencv-python